Change temp type to uint32_t to resolve gcc warning.
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/6412


git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@24378 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/common/ucnvisci.c
@@ -1007,7 +1007,7 @@ static void UConverter_fromUnicode_ISCII_OFFSETS_LOGIC(
                     /* we are in a script block which is different than
                      * previous sourceChar's script block write ATR and language codes
                      */
-                    uint16_t temp=0;
+                    uint32_t temp=0;
                     temp =(uint16_t)(ATR<<8);
                     temp += (uint16_t)((uint8_t) lookupInitialData[range].isciiLang);
                     /* reset */
